Update and rename Yohan to Yohan.java

+++ 1_입출력과 사칙연산/8_A÷B/Yohan.java
@@ -3,8 +3,8 @@ public class Main{
     public static void main(String[] args){
         double a, b;
         Scanner sc = new Scanner(System.in);
-        a = sc.nextInt();
-        b = sc.nextInt();
+        a = sc.nextDouble();
+        b = sc.nextDouble();
         System.out.println(a/b);
         
     }
